requirements.txt
anthropic
streamlit
lancedb
pymupdf
pandas
python-dotenv
